Move material selection into overridable formulas
